% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alder.R
\name{age}
\alias{age}
\title{Calculate age}
\usage{
age(dob, age.day = today(), units = "years", floor = TRUE)
}
\arguments{
\item{dob}{date-of-birth, the day to start calculating age.}

\item{age.day}{the date on which age is to be calculated.}

\item{units}{unit to measure age in. Defaults to "years". Passed to duration.}

\item{floor}{boolean for whether or not to floor the result. Defaults to TRUE.}
}
\value{
Age in units. Will be an integer if floor = TRUE.
}
\description{
By default, calculates the typical "age in years", with a
floor applied so that you are, e.g., 5 years old from
5th birthday through the day before your 6th birthday. Set
\code{floor = FALSE} to return decimal ages, and change units
for units other than years.
}
\examples{
my.dob <- as.Date('1983-10-20')
age(my.dob)
age(my.dob, units = "minutes")
age(my.dob, floor = FALSE)

}
