% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesogprosesser.R
\name{lesogprosesser}
\alias{lesogprosesser}
\title{Les og preprosesser data for bruk i Norvas sine rapporter}
\usage{
lesogprosesser()
}
\value{
Ferdig prosessert data for Norvas
}
\description{
Denne funksjonen gjor nodvendig preprosessering av Norvas sin data for bruk i rapporter
}
