% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasFigAndelGruppert.R
\name{norvasFigAndelGruppert}
\alias{norvasFigAndelGruppert}
\title{Lag gruppert andelsplot}
\usage{
norvasFigAndelGruppert(
  RegData,
  var = NA,
  grvar = NA,
  tittel = "",
  skriftStr = 1.3,
  pktStr = 1.4,
  decreasing = F,
  outfile = "",
  width = 800,
  height = 700
)
}
\arguments{
\item{RegData}{En dataramme med kolonne Sykehusnavn, Aar, Teller og Nevner}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}
}
\value{
En figur med søylediagram over valgt variabel
}
\description{
Denne funksjonen lager et søylediagram over valgt dikotom variabel
}
