% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasFigAndelTid.R
\name{norvasFigAndelTid}
\alias{norvasFigAndelTid}
\title{Tidstrend (år/mnd) av rate/andel for en gitt variabel}
\usage{
norvasFigAndelTid(RegData = 0, valgtVar = "ErMann",
  datoFra = "2014-01-01", datoTil = "2050-12-31", minald = 0,
  maxald = 130, erMann = 99, outfile = "",
  datovar = "InklusjonDato", reshID, enhetsUtvalg = 1,
  inkl_konf = 99, aldervar = "PatientAge", valgtShus = c(""),
  tidsenhet = "Aar", diag_gruppe = 99)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{inkl_konf}{Inkluder konfidensintervall i figur
0: Nei
1: Ja
99: Bruk default som definert i NorgastPrepVar}

\item{valgtShus}{Vektor med UnitId over hvilke sykehus man genererer rapporten for.
Denne overstyrer reshID og er bare tilgjengelig for SC-bruker.}

\item{tidsenhet}{Plot figur med år eller måned som tidsenhet
'Aar' (Default)
'Mnd'}
}
\value{
En figur med tidsutvikling av rate over år
}
\description{
Årlige eller månedlige rater for valgt variabel.
Konfidensintervall kan inkluderes hvis ønskelig.
}
\details{
Konfidensintervallet er basert på Clopper Pearsons "eksakte" metode for binominalfordelt data.
}
