% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasIndikator.R
\name{norvasIndikator}
\alias{norvasIndikator}
\title{Lag indikatorfigur}
\usage{
norvasIndikator(
  indikatordata,
  tittel = "",
  terskel = 5,
  minstekrav = NA,
  maal = NA,
  skriftStr = 1.3,
  pktStr = 1.4,
  legPlass = "top",
  minstekravTxt = "Min.",
  maalTxt = "Mål",
  graaUt = NA,
  decreasing = F,
  outfile = "",
  lavDG = NA,
  width = 800,
  height = 700,
  inkl_konf = F,
  maalretn = "hoy",
  xmax = NA
)
}
\arguments{
\item{indikatordata}{En dataramme med kolonnene Sykehusnavn, Aar, Teller og Nevner}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}
}
\value{
En figur med søylediagram over valgt indikator
}
\description{
Denne funksjonen lager et søylediagram over valgt indikator
}
