% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasIndikatorOppf.R
\name{norvasIndikatorOppf}
\alias{norvasIndikatorOppf}
\title{Lag søylediagram eller stabelplott som viser andeler av ulike variabler}
\usage{
norvasIndikatorOppf(
  RegData = 0,
  valgtVar = "Inklusjonsalder",
  datoFra = "2014-01-01",
  datoTil = "2050-12-31",
  minald = 0,
  maxald = 130,
  erMann = 99,
  outfile = "",
  reshID,
  enhetsUtvalg = 0,
  preprosess = F,
  valgtShus = c(""),
  hentData = F,
  datovar = "InklusjonDato",
  aldervar = "PatientAge",
  diag_gruppe = 99
)
}
\value{
En figur med søylediagram eller et stabelplot av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram eller stabelplot som viser andeler av valgt variabel
filtrert på de utvalg som er gjort.
}
