% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasKodebok.R
\name{norvasKodebok}
\alias{norvasKodebok}
\title{Lag og dokumenter kodebok}
\usage{
norvasKodebok()
}
\value{
En formatert kodebok
}
\description{
Denne funksjonen leser kodebok i xlsx-format og lager kodebok i Rdata-format.
}
