% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasPreprosess.R
\name{norvasPreprosess}
\alias{norvasPreprosess}
\title{Preprosesser data for bruk i Norvas sine rapporter}
\usage{
norvasPreprosess(RegData)
}
\arguments{
\item{Regdata}{En dataramme med registerdata}
}
\value{
Et preprosessert datasett
}
\description{
Denne funksjonen gjor nodvendig preprosessering av Norvas sin data for bruk i rapporter
}
