% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasPrepvar.R
\name{norvasPrepvar}
\alias{norvasPrepvar}
\title{Klargjør variabler for fremstilling i figur eller tabell}
\usage{
norvasPrepvar(RegData, valgtVar)
}
\arguments{
\item{valgtVar}{Variabelen som skal fremstilles}
}
\value{
PrepData En liste med plotrelevante størrelser
}
\description{
Klargjør variabler for fremstilling i figur eller tabell
}
