% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasStabelGrvar.R
\name{norvasStabelGrvar_v2}
\alias{norvasStabelGrvar_v2}
\title{Plot stablede andeler per grupperingsvariabel}
\usage{
norvasStabelGrvar_v2(
  plotMatrise,
  grtxt = "",
  outfile = "",
  legendTxt = NA,
  tittel = "",
  xlab = " Antall pasienter",
  cexgr = 1,
  fargepalett = "BlaaOff",
  beside = F,
  revcol = F,
  revcol_legend = F,
  pstTxt = NA,
  NTxt = NA
)
}
\description{
Stabelplot, endret legendplassering
}
