% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasUtvalg.R
\name{norvasUtvalg}
\alias{norvasUtvalg}
\title{Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.}
\usage{
norvasUtvalg(RegData, datoFra, datoTil, datovar = "InklusjonDato",
  minald, maxald, aldervar = "PatientAge", erMann, valgtShus = "",
  fargepalett = "BlaaRapp", diag_gruppe = 99)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{valgtShus}{Vektor med UnitId over hvilke sykehus man genererer rapporten for.
Denne overstyrer reshID og er bare tilgjengelig for SC-bruker.}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.
}
