% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnr2fodselsdato.R
\name{personnr2fodselsdato}
\alias{personnr2fodselsdato}
\title{Trekk ut fødselsdato fra personnummer}
\usage{
personnr2fodselsdato(pnr)
}
\arguments{
\item{pnr}{Norsk fødselsnummer enten som skalar eller vektor}
}
\value{
fodselsdato Fødselsdato i datoformat
}
\description{
Trekk ut fødselsdato fra personnummer
}
