% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner.R
\name{wrap.it}
\alias{wrap.it}
\title{Automatisk linjebryting av lange tekstetiketter}
\usage{
wrap.it(x, len)
}
\arguments{
\item{x}{En tekststreng eller vektor av tekststrenger}

\item{len}{Lengden strengen skal brytes ved}
}
\value{
En tekststreng med tekstbrudd på angitt lengde
}
\description{
Denne funksjonen tar som input en vektor med streng-elementer og returnerer
en vektor med \n satt inn for linjebrudd
}
