% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasSabelGrvar.R
\name{norvarStabelGrvar}
\alias{norvarStabelGrvar}
\title{Plot stablede andeler per grupperingsvariabel}
\usage{
norvarStabelGrvar(
  plotMatrise,
  grtxt = "",
  outfile = "",
  legendTxt = NA,
  tittel = "",
  xlab = " Antall pasienter",
  cexgr = 1,
  fargepalett = "BlaaOff"
)
}
\description{
Stabelplot
}
