% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norvasFigAndelGruppert.R
\name{norvasFigAndelGruppert}
\alias{norvasFigAndelGruppert}
\title{Lag gruppert andelsplot}
\usage{
norvasFigAndelGruppert(
  RegData,
  var = NA,
  grvar = NA,
  tittel = "",
  skriftStr = 1.3,
  decreasing = F,
  outfile = "",
  width = 800,
  height = 700,
  mrom = 0.3,
  datoFra = "2000-01-01",
  datoTil = "2050-01-01",
  datovar = "InklusjonDato",
  minald = 0,
  maxald = 130,
  aldervar = "PatientAge"
)
}
\arguments{
\item{RegData}{En dataramme med kolonne Sykehusnavn, Aar, Teller og Nevner}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}
}
\value{
En figur med søylediagram over valgt variabel
}
\description{
Denne funksjonen lager et søylediagram over valgt dikotom variabel
}
